workers/requirements.txt
openai>=1.0.0
redis>=5.0.0
pymupdf>=1.23.0
Pillow>=10.0.0
httpx>=0.25.0
tabulate>=0.9.0
